refactor: 로또번호 일급컬렉션으로 변경
