+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -7,34 +7,13 @@ import java.util.List;
 
 public class Lotto {
 
-    private final List<Integer> numbers;
+    private final LottoNumbers lottoNumbers;
 
     private Lotto(GenerateStrategy strategy) {
-        this.numbers = validate(strategy.generate());
-        ;
+        this.lottoNumbers = LottoNumbers.of(strategy.generate());
     }
 
-    private List<Integer> validate(List<Integer> numbers) {
-        validateCount(numbers);
 
-        for (Integer number : numbers) {
-            validateRange(number);
-        }
-
-        return numbers;
-    }
-
-    private void validateCount(List<Integer> numbers) {
-        if (numbers.size() != LottoConstants.NUMBER_COUNT) {
-            throw new IllegalArgumentException(LottoConstants.NUMBER_COUNT_ERROR_MESSAGE);
-        }
-    }
-
-    private void validateRange(Integer number) {
-        if (number < LottoConstants.NUMBER_MIN || number > LottoConstants.NUMBER_MAX) {
-            throw new IllegalArgumentException(LottoConstants.NUMBER_ERROR_MESSAGE);
-        }
-    }
 
     public static Lotto from(GenerateStrategy strategy) {
         return new Lotto(strategy);
